refactor: vue-router navigation guard
